Don't delete target non-regular files in SCRIPT TO

+++ h2/src/main/org/h2/command/dml/ScriptBase.java
@@ -104,7 +104,9 @@ abstract class ScriptBase extends Prepared {
     void deleteStore() {
         String file = getFileName();
         if (file != null) {
-            FileUtils.delete(file);
+            if (FileUtils.isRegularFile(file)) {
+                FileUtils.delete(file);
+            }
         }
     }
